+++ AutoMate/XCTest extensions/XCUIElement+Swipe.swift
@@ -28,8 +28,8 @@ extension XCUIElement {
     ///   - startVector: Relative point from which to start swipe.
     ///   - stopVector: Relative point to end swipe.
     public func swipe(from startVector: CGVector, to stopVector: CGVector) {
-        let p1 = coordinate(withNormalizedOffset: startVector)
-        let p2 = coordinate(withNormalizedOffset: stopVector)
+        let p1 = smartCoordinate(withNormalizedOffset: startVector)
+        let p2 = smartCoordinate(withNormalizedOffset: stopVector)
         p1.press(forDuration: 0.1, thenDragTo: p2)
     }